fix: add max-width clamp to ReviewPage file browser + extract shared FileBrowserContent
- ReviewPage: mirror IssueDetailPage's max-width clamping logic to prevent
  file browser from collapsing the chat area below MIN_CHAT_WIDTH
- Extract FileBrowserContent shared component to deduplicate header,
  breadcrumb, and content area between FileBrowserDrawer and FileBrowserPanel
